+++ db/migrate/20220212232000_add_assigned_user_id_to_claims.rb
@@ -1,0 +1,5 @@
+class AddAssignedUserIdToClaims < ActiveRecord::Migration[6.0]
+  def change
+    add_column :claims, :assigned_to_id, :string
+  end
+end