+++ spec/lib/claim_allocator_spec.rb
@@ -1,0 +1,30 @@
+require "rails_helper"
+require "claim_allocator"
+
+RSpec.describe ClaimAllocator do
+  let(:ecp_claim) { create(:claim, :submitted, policy: EarlyCareerPayments) }
+  let(:tslr_claim) { create(:claim, :submitted, policy: StudentLoans) }
+  let(:maxime) { create(:dfe_signin_user, given_name: "Maxime", family_name: "Toussaint", organisation_name: "Department for Education", role_codes: [DfeSignIn::User::SERVICE_OPERATOR_DFE_SIGN_IN_ROLE_CODE]) }
+  let(:sofia) { create(:dfe_signin_user, given_name: "Sofia", family_name: "Bianchi", organisation_name: "Department for Education", role_codes: [DfeSignIn::User::SERVICE_OPERATOR_DFE_SIGN_IN_ROLE_CODE]) }
+
+  it "assigns a claim team member to a claim" do
+    expect(ecp_claim.assigned_to_id).to be_nil
+
+    ClaimAllocator.new(claim_ids: ecp_claim.id, admin_user_id: sofia.id).call
+
+    expect(ecp_claim.reload.assigned_to).not_to be_nil
+    expect(ecp_claim.assigned_to.full_name).to eq "Sofia Bianchi"
+  end
+
+  it "a claim team member can be assign mutiple claims" do
+    expect(ecp_claim.assigned_to_id).to be_nil
+    expect(tslr_claim.assigned_to_id).to be_nil
+
+    ClaimAllocator.new(claim_ids: [ecp_claim.id, tslr_claim.id], admin_user_id: maxime.id).call
+
+    expect(ecp_claim.reload.assigned_to).not_to be_nil
+    expect(tslr_claim.reload.assigned_to).not_to be_nil
+    expect(ecp_claim.assigned_to.full_name).to eq "Maxime Toussaint"
+    expect(tslr_claim.assigned_to.full_name).to eq "Maxime Toussaint"
+  end
+end